Move Java Toolchains integration tests to :toolchains-jvm
