+++ platforms/jvm/toolchains-jvm/build.gradle.kts
@@ -61,6 +61,7 @@ dependencies {
     testImplementation(testFixtures(projects.core))
     testImplementation(testFixtures(projects.logging))
     testImplementation(testFixtures(projects.toolchainsJvmShared))
+    testImplementation(testFixtures(projects.languageJava))
 
     testFixturesImplementation(projects.native)
