+++ platforms/jvm/toolchains-jvm/src/integTest/groovy/org/gradle/jvm/toolchain/JavaToolchainBuildOperationsIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 the original author or authors.
+ * Copyright 2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.api.tasks
+package org.gradle.jvm.toolchain
 
+import org.gradle.api.tasks.JavaExecToolchainFixture
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.AvailableJavaHomes
 import org.gradle.integtests.fixtures.executer.GradleContextualExecuter
