+++ platforms/jvm/toolchains-jvm/src/integTest/groovy/org/gradle/jvm/toolchain/JavaToolchainDownloadSpiKotlinIntegrationTest.groovy
